refactor: FullHeightButtonGroup not last-child에 우측 border 0 설정, 공용 component로 이동
- FullHeightButtonGroup 스타일 컴포넌트에 not last-child일 경우 우측 border  0을 important로 설정

- FullHeightButtonGroup을 공용 component/styled.tsx로 이동
